Experiment new pattern on ReturnStatement.
