+++ compiler/ast/test_helper_types.go
@@ -13,17 +13,6 @@ type testingNode interface {
 	NameIs(name string) bool
 }
 
-type TestingStatement interface {
-	Statement
-	// Test Helpers
-	IsClassStmt(t *testing.T, className string) *ClassStatement
-	IsDefStmt(t *testing.T) *DefStatement
-	IsExpression(t *testing.T) TestingExpression
-	IsModuleStmt(t *testing.T, className string) *ModuleStatement
-	IsReturnStmt(t *testing.T) *ReturnStatement
-	IsWhileStmt(t *testing.T) *WhileStatement
-}
-
 type TestingExpression interface {
 	Expression
 	// Test Helpers
@@ -39,10 +28,4 @@ type TestingExpression interface {
 	IsIntegerLiteral(t *testing.T) *IntegerLiteral
 	IsStringLiteral(t *testing.T) *StringLiteral
 	IsYieldExpression(t *testing.T) *YieldExpression
-}
-
-type CodeBlock []TestingStatement
-
-func (cb CodeBlock) NthStmt(n int) TestingStatement {
-	return cb[n-1]
 }
